refactor(racingcar): if문을 when으로, 상수화, Constants파일 위치이동
코틀린 컨벤션에 맞춰 3가지 이상의 옵션이 있는 if문을 when문으로 변경했다. 변경 오류의 위험과, 명확성, 컴파일 시점에 할당하기 위해 상수화되지 않은 숫자들을 상수화했다.
