+++ src/main/kotlin/racingcar/domain/CarName.kt
@@ -1,7 +1,6 @@
 package racingcar.domain
 
 import camp.nextstep.edu.missionutils.Console.readLine
-import racingcar.CarNameConstants
 
 class CarName {
     fun requestInput() {
@@ -13,30 +12,17 @@ class CarName {
     }
 
     fun validateInput(input: String): List<String> {
-        if (input.isEmpty()) {
-            throw IllegalArgumentException(CarNameConstants.INPUT_IS_EMPTY)
-        }
-
-        if (!input.contains(CarNameConstants.COMMA_DELIMITER)) {
-            throw IllegalArgumentException(CarNameConstants.INPUT_ONLY_ONE)
+        when {
+            input.isEmpty() -> throw IllegalArgumentException(CarNameConstants.INPUT_IS_EMPTY)
+            !input.contains(CarNameConstants.COMMA_DELIMITER) -> throw IllegalArgumentException(CarNameConstants.INPUT_ONLY_ONE)
         }
 
         val splitInput = input.split(CarNameConstants.COMMA_DELIMITER)
-
-        if (splitInput.any { it.isEmpty() }) {
-            throw IllegalArgumentException(CarNameConstants.CAR_NAME_HAS_EMPTY)
-        }
-
-        if (splitInput.any { it.isBlank() }) {
-            throw IllegalArgumentException(CarNameConstants.CAR_NAME_HAS_BLANK)
-        }
-
-        if (splitInput.any { it.length > 5 }) {
-            throw IllegalArgumentException(CarNameConstants.CAR_NAME_HAS_OVER_5_CHARS)
-        }
-
-        if (splitInput.toSet().size != splitInput.size) {
-            throw IllegalArgumentException(CarNameConstants.CAR_NAME_HAS_DUPLICATE)
+        when {
+            splitInput.any { it.isEmpty() } -> throw IllegalArgumentException(CarNameConstants.CAR_NAME_HAS_EMPTY)
+            splitInput.any { it.isBlank() } -> throw IllegalArgumentException(CarNameConstants.CAR_NAME_HAS_BLANK)
+            splitInput.any { it.length > CarNameConstants.CAR_NAME_MAX_LENGTH } -> throw IllegalArgumentException(CarNameConstants.CAR_NAME_HAS_OVER_5_CHARS)
+            splitInput.toSet().size != splitInput.size -> throw IllegalArgumentException(CarNameConstants.CAR_NAME_HAS_DUPLICATE)
         }
 
         return splitInput
