+++ src/main/kotlin/racingcar/domain/CarNameConstants.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 object CarNameConstants {
     const val REQUEST_INPUT = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)"
@@ -7,6 +7,7 @@ object CarNameConstants {
     const val INPUT_ONLY_ONE = "자동차 이름이 1대만 입력되었습니다."
     const val CAR_NAME_HAS_EMPTY = "자동차 이름 중 빈값이 있습니다."
     const val CAR_NAME_HAS_BLANK = "자동차 이름 중 공백이 있습니다."
+    const val CAR_NAME_MAX_LENGTH = 5
     const val CAR_NAME_HAS_OVER_5_CHARS = "자동차 이름 중 5자가 초과된 값이 있습니다."
     const val CAR_NAME_HAS_DUPLICATE = "자동차 이름 중 중복된 이름이 있습니다."
 }
