+++ src/main/kotlin/racingcar/domain/CarRaceConstants.kt
@@ -1,6 +1,8 @@
-package racingcar
+package racingcar.domain
 
 object CarRaceConstants {
+    const val RANDOM_NUMBER_MIN = 0
+    const val RANDOM_NUMBER_MAX = 10
     const val INITIALIZATION_ZERO = 0
     const val SCORE_ONE_POINT = 1
     const val SCORE_CUT_LINE = 4
