+++ src/main/kotlin/racingcar/domain/RefereeConstants.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 object RefereeConstants {
     const val WINNER_MESSAGE = "최종 우승자 : "
